+++ src/shared/index.js
@@ -15,7 +15,7 @@ export { default as useConnectionStatus } from './use-connection-status.js'
 export * from './use-context-selection/index.js'
 export * from './use-data-value-set/index.js'
 export * from './use-is-first-render.js'
-export * from './use-server-time-as-local/index.js'
+export * from './use-now-at-server-timezone/index.js'
 export * from './use-user-info/index.js'
 export * from './utils.js'
 export * from './validation/index.js'
