--- src/shared/use-server-time-as-local/index.js
@@ -1,2 +1,0 @@
-export * from './use-now-at-server-timezone.js'
-export * from './use-server-date-time-as-local.js'
